chat-with-persistent-pdf/requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
pdfplumber
chromadb
